update: api RESAS| 人口構成api関連ファイルを populationCompositionフォルダにまとめる
