+++ src/api/RESAS/populationComposition/mock.ts
@@ -1,7 +1,4 @@
-import type {
-  YearPopulation,
-  PopulationCompositionPerYear,
-} from "../populationComposition";
+import type { YearPopulation, PopulationCompositionPerYear } from "./type";
 
 const MAX_POPULATION = 10 ** 9; // NOTE: 東京の2045年は8桁代だったため、この値が入るように
 const DECIMAL_PLACE_RATE = 100;
